+++ App/xcbuddykit/Sources/Extensions/ArgumentParserError+FatalError.swift
@@ -1,0 +1,9 @@
+import Foundation
+import Utility
+
+extension ArgumentParserError: FatalError {
+    /// Error type
+    var type: ErrorType {
+        return .abort
+    }
+}