+++ App/xcbuddykit/Sources/Generator/GeneratorError.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-enum GeneratorError: Error, ErrorStringConvertible {
-    var errorDescription: String {
-        return ""
-    }
-}