+++ App/xcbuddykit/Sources/Loader/GraphLoaderContext.swift
@@ -33,18 +33,16 @@ class GraphLoaderContext: Context, GraphLoaderContexting {
     ///   - circularDetector: Circular dependency detector.
     ///   - shell: shell.
     ///   - printer: printer.
-    ///   - errorHandler: error handler.
     init(manifestLoader: GraphManifestLoading = GraphManifestLoader(),
          cache: GraphLoaderCaching = GraphLoaderCache(),
          fileHandler: FileHandling = FileHandler(),
          circularDetector: GraphCircularDetecting = GraphCircularDetector(),
          shell: Shelling = Shell(),
-         printer: Printing = Printer(),
-         errorHandler: ErrorHandling = ErrorHandler()) {
+         printer: Printing = Printer()) {
         self.manifestLoader = manifestLoader
         self.cache = cache
         self.circularDetector = circularDetector
-        super.init(fileHandler: fileHandler, shell: shell, printer: printer, errorHandler: errorHandler)
+        super.init(fileHandler: fileHandler, shell: shell, printer: printer)
     }
 
     /// Initializes the graph loader context with a context and the extra attributes that the graph loader context has.
@@ -61,6 +59,6 @@ class GraphLoaderContext: Context, GraphLoaderContexting {
         self.manifestLoader = manifestLoader
         self.cache = cache
         self.circularDetector = circularDetector
-        super.init(fileHandler: context.fileHandler, shell: context.shell, printer: context.printer, errorHandler: context.errorHandler)
+        super.init(fileHandler: context.fileHandler, shell: context.shell, printer: context.printer)
     }
 }