+++ App/xcbuddykit/Sources/Utils/Context.swift
@@ -13,9 +13,6 @@ protocol Contexting: AnyObject {
 
     /// Resource locator.
     var resourceLocator: ResourceLocating { get }
-
-    /// Error handler.
-    var errorHandler: ErrorHandling { get }
 }
 
 /// xcbuddy uses contexts as a dependency injection mechanism.
@@ -33,9 +30,6 @@ class Context: Contexting {
     /// Resource locator.
     let resourceLocator: ResourceLocating
 
-    /// Error handler.
-    let errorHandler: ErrorHandling
-
     /// Initializes the context with its attributess.
     ///
     /// - Parameters:
@@ -46,12 +40,10 @@ class Context: Contexting {
     init(fileHandler: FileHandling = FileHandler(),
          shell: Shelling = Shell(),
          printer: Printing = Printer(),
-         resourceLocator: ResourceLocating = ResourceLocator(),
-         errorHandler: ErrorHandling = ErrorHandler()) {
+         resourceLocator: ResourceLocating = ResourceLocator()) {
         self.fileHandler = fileHandler
         self.shell = shell
         self.printer = printer
         self.resourceLocator = resourceLocator
-        self.errorHandler = errorHandler
     }
 }