+++ Tests/ProjectDescriptionTests/TargetTests.swift
@@ -7,6 +7,7 @@ final class TargetTests: XCTestCase {
         let subject = Target(name: "name",
                              platform: .ios,
                              product: .app,
+                             bundleId: "bundle_id",
                              infoPlist: "info.plist",
                              entitlements: "entitlements",
                              dependencies: [.framework(path: "path")],
@@ -14,9 +15,7 @@ final class TargetTests: XCTestCase {
                                                 release: .settings([:], xcconfig: "release.xcconfig")),
                              buildPhases: [.headers()])
         let json = subject.toJSON()
-        let expected = """
-        {"build_phases": [{"private": [], "project": [], "public": [], "type": "headers"}], "dependencies": [{"path": "path", "type": "framework"}], "entitlements": "entitlements", "info_plist": "info.plist", "name": "name", "platform": "ios", "product": "app", "settings": {"base": {}, "debug": {"settings": {}, "xcconfig": "debug.xcconfig"}, "release": {"settings": {}, "xcconfig": "release.xcconfig"}}}
-        """
+        let expected = "{\"build_phases\": [{\"private\": [], \"project\": [], \"public\": [], \"type\": \"headers\"}], \"bundle_id\": \"bundle_id\", \"dependencies\": [{\"path\": \"path\", \"type\": \"framework\"}], \"entitlements\": \"entitlements\", \"info_plist\": \"info.plist\", \"name\": \"name\", \"platform\": \"ios\", \"product\": \"app\", \"settings\": {\"base\": {}, \"debug\": {\"settings\": {}, \"xcconfig\": \"debug.xcconfig\"}, \"release\": {\"settings\": {}, \"xcconfig\": \"release.xcconfig\"}}}"
         XCTAssertEqual(json.toString(), expected)
     }
 }